requirements.txt
discord.py>=2.7.0
davey
pynacl
python-dotenv
requests